feat: add method for db

+++ src/internal/logic/campaigns/domain/models/campaigns.go
@@ -6,7 +6,7 @@ type Campaigns struct {
 	StartDate   string  `gorm:"column:start_date" json:"startDate"`
 	EndDate     string  `gorm:"column:end_date" json:"endDate"`
 	BonusType   string  `gorm:"column:bonus_type" json:"bonusType"`
-	BonusValue  float32 `gorm:"column:bonus_value" json:"bonusValue"`
+	BonusValue  float64 `gorm:"column:bonus_value" json:"bonusValue"`
 	MinPurchase string  `gorm:"column:min_purchase" json:"minPurchase"`
 	CreatedAt   string  `gorm:"column:created_at" json:"createdAt"`
 }
